--- Hackathon Master Folder/HackWeb.html
@@ -1,29 +1,0 @@
-<!doctype html>
-<html lang="en">
-    <!-- Need this to use everything on Bootstrap: -->
-    <head>
-        <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha2/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-aFq/bzH65dt+w6FI2ooMVUpc+21e0SRygnTpmBvdBgSdnuTN7QbdgL+OapgHtvPp" crossorigin="anonymous">
-        <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha2/dist/js/bootstrap.bundle.min.js" integrity="sha384-qKXV1j0HvMUeCBQ+QVp7JcfGl760yU08IQ+GpUo5hlbpg51QRiuqHAJz8+BrxE/N" crossorigin="anonymous"></script> 
-        <script src="https://cdn.jsdelivr.net/npm/@popperjs/core@2.11.6/dist/umd/popper.min.js" integrity="sha384-oBqDVmMz9ATKxIep9tiCxS/Z9fNfEXiDAYTujMAeBAsjFuCZSmKbSSUnQlmh/jp3" crossorigin="anonymous"></script>
-        <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha2/dist/js/bootstrap.min.js" integrity="sha384-heAjqF+bCxXpCWLa6Zhcp4fu20XoNIA98ecBC1YkdXhszjoejr5y9Q77hIrv8R9i" crossorigin="anonymous"></script>
-    </head>
-    <title> Green Financier</title>
-    <link rel="icon" type="image/x-icon" href="assets/fav.png">
-    <body>
-        <!-- This is where everything should be copy-pasted -->
-        <!-- Found under Components->Navbar (edited): -->
-        <!--  -->
-  
-            <div class="row g-0 bg-light position-relative">
-              <div class="col-md-3 mb-md-0 p-md-4">
-                <img src="assets/fav.png" class="img-fluid" width="100%" height="250">    
-
-              </div>
-              <div class="col-md-9 p-4 ps-md-0">
-                <h3 class="mt-0">Empower your investments for a better planet.</h3>
-              </div>
-            </div>
-              
-    
-    </body>
-</html>
